Make InstrumentationLibrary<signal>ToScope functions unexported
The functions were added in the the latest version. They are intended for the internal use only and shouldn't be exported.
