+++ model/otlpgrpc/logs.go
@@ -23,8 +23,8 @@ import (
 
 	ipdata "go.opentelemetry.io/collector/model/internal"
 	otlpcollectorlog "go.opentelemetry.io/collector/model/internal/data/protogen/collector/logs/v1"
-	v1 "go.opentelemetry.io/collector/model/internal/data/protogen/common/v1"
 	otlplogs "go.opentelemetry.io/collector/model/internal/data/protogen/logs/v1"
+	"go.opentelemetry.io/collector/model/internal/otlp"
 	"go.opentelemetry.io/collector/model/pdata"
 )
 
@@ -85,7 +85,7 @@ func (lr LogsRequest) UnmarshalProto(data []byte) error {
 	if err := lr.orig.Unmarshal(data); err != nil {
 		return err
 	}
-	InstrumentationLibraryLogsToScope(lr.orig.ResourceLogs)
+	otlp.InstrumentationLibraryLogsToScope(lr.orig.ResourceLogs)
 	return nil
 }
 
@@ -103,7 +103,7 @@ func (lr LogsRequest) UnmarshalJSON(data []byte) error {
 	if err := jsonUnmarshaler.Unmarshal(bytes.NewReader(data), lr.orig); err != nil {
 		return err
 	}
-	InstrumentationLibraryLogsToScope(lr.orig.ResourceLogs)
+	otlp.InstrumentationLibraryLogsToScope(lr.orig.ResourceLogs)
 	return nil
 }
 
@@ -161,30 +161,4 @@ type rawLogsServer struct {
 func (s rawLogsServer) Export(ctx context.Context, request *otlpcollectorlog.ExportLogsServiceRequest) (*otlpcollectorlog.ExportLogsServiceResponse, error) {
 	rsp, err := s.srv.Export(ctx, LogsRequest{orig: request})
 	return rsp.orig, err
-}
-
-// InstrumentationLibraryLogsToScope implements the translation of resource logs data
-// following the v0.15.0 upgrade:
-//      receivers SHOULD check if instrumentation_library_logs is set
-//      and scope_logs is not set then the value in instrumentation_library_logs
-//      SHOULD be used instead by converting InstrumentationLibraryLogs into ScopeLogs.
-//      If scope_logs is set then instrumentation_library_logs SHOULD be ignored.
-// https://github.com/open-telemetry/opentelemetry-proto/blob/3c2915c01a9fb37abfc0415ec71247c4978386b0/opentelemetry/proto/logs/v1/logs.proto#L58
-func InstrumentationLibraryLogsToScope(rls []*otlplogs.ResourceLogs) {
-	for _, rl := range rls {
-		if len(rl.ScopeLogs) == 0 {
-			for _, ill := range rl.InstrumentationLibraryLogs {
-				scopeLogs := otlplogs.ScopeLogs{
-					Scope: v1.InstrumentationScope{
-						Name:    ill.InstrumentationLibrary.Name,
-						Version: ill.InstrumentationLibrary.Version,
-					},
-					LogRecords: ill.LogRecords,
-					SchemaUrl:  ill.SchemaUrl,
-				}
-				rl.ScopeLogs = append(rl.ScopeLogs, &scopeLogs)
-			}
-		}
-		rl.InstrumentationLibraryLogs = nil
-	}
 }
